Fixed includes so that std::abs works with integers.

+++ assignment2/arcade.cpp
@@ -29,7 +29,7 @@
 #include <iostream>
 #include <limits>
 #include <string>
-#include <cmath>
+#include <cstdlib>
 using namespace std;
 
 // Function prototypes
@@ -48,7 +48,7 @@ int main()
     // Because this is more fun than reprompting
     if (coupons < 0)
     {
-        cout << endl << "You are " << abs(static_cast<float>(coupons))
+        cout << endl << "You are " << abs(coupons)
              << " coupon" << (coupons == -1 ? "" : "s") << " in debt." << endl
              << "You owe the arcade prizes!" << endl;
     }
